Prevent serialization errors from affecting UI
Previously, if concurrent requests caused a serialization error (a
transaction failed to commit), this was made visible in the UI, as an
internal server error, and the user had to manually re-try the
operation. This was a bad user experience.

This patch changes Gavel's functionality to avoid exposing this error to
the user and instead handle serialization errors internally by re-trying
transactions until they succeed.