Resolve XSS vuln: do not mark unsafe inputs as 'safe'
This commit resolves an XSS vulnerability in Gavel. It is Jinja's default
behaviour to escape any unsafe inputs so that they can be inserted into HTML
content without leading to a security vulnerability. However, previous versions
of Gavel used the "safe" filter to override this. As a result, if a malicious
participant entered a description, email, etc. which contained HTML code, anybody
who viewed the payload could have their session with Gavel hijacked if the HTML
contained malicious JavaScript code. This class of vulnerability is known as
'cross-site scripting'.

To resolve the vulnerability, all instances of the "safe" filter are removed. Hackathon
organisers on older versions should sanitise inputs before importing them into Gavel.